remove calls to GeneralPath method that was added in Java 1.6

+++ trunk/simulations-java/simulations/line-graphing/src/edu/colorado/phet/linegraphing/intro/view/RiseRunBracketNode.java
@@ -138,13 +138,13 @@ class RiseRunBracketNode extends PComposite {
     private static Shape createBracketShape( double width ) {
         final double bracketTipWidth = Math.min( BRACKET_TIP_WIDTH, 0.5 * width );
         GeneralPath path = new GeneralPath();
-        path.moveTo( 0, BRACKET_END_Y_OFFSET );
-        path.quadTo( 0, (float) BRACKET_END_HEIGHT, (float) BRACKET_CORNER_RADIUS, (float) BRACKET_END_HEIGHT );
+        path.moveTo( 0f, (float) BRACKET_END_Y_OFFSET );
+        path.quadTo( 0f, (float) BRACKET_END_HEIGHT, (float) BRACKET_CORNER_RADIUS, (float) BRACKET_END_HEIGHT );
         path.lineTo( (float) ( ( width - bracketTipWidth ) / 2 ), (float) BRACKET_END_HEIGHT );
         path.lineTo( (float) ( width / 2 ), (float) ( BRACKET_END_HEIGHT + BRACKET_TIP_HEIGHT ) );
         path.lineTo( (float) ( ( width + bracketTipWidth ) / 2 ), (float) BRACKET_END_HEIGHT );
         path.lineTo( (float) ( width - BRACKET_CORNER_RADIUS ), (float) BRACKET_END_HEIGHT );
-        path.quadTo( (float) width, (float) BRACKET_END_HEIGHT, (float) width, BRACKET_END_Y_OFFSET );
+        path.quadTo( (float) width, (float) BRACKET_END_HEIGHT, (float) width, (float) BRACKET_END_Y_OFFSET );
         return path;
     }
